Fix 5 integration issues: DBSQL VM costs, Model Serving dropdown, FMAPI export, Lakebase sizes
Sprint A: DBSQL Classic/Pro export — add DB-backed VM cost lookup via warehouse
config table (sync_ref_dbsql_warehouse_config → sync_pricing_vm_costs). Pass db
session through export builder pipeline. Use dbsql_vm_pricing_tier for display.

Sprint B: Model Serving GPU/CPU dropdown — transform API response field names
(gpu_type→id, dbu_rate→dbu_per_hour) in store's fetchModelServingGPUTypes.

Sprint C: FMAPI Proprietary export — fix context_length default from 'long' to
'all' to match API calculation (fmapi_calc.py). Affects rate lookups for Google
models that only have 'long'/'short' contexts.

Sprint D: FMAPI Databricks Azure — verified pricing data exists and export path
is correct. No code changes needed.

Sprint E: Lakebase CU sizes — update frontend dropdown from old 4 options (1,2,4,8)
to full range (0.5–112 CU) with Autoscale/Fixed groups. Fix parseInt→parseFloat
for 0.5 CU support. Update model column type to Numeric(5,1).

+++ backend/app/models/line_item.py
@@ -57,7 +57,7 @@ class LineItem(Base):
     fmapi_quantity = Column(Numeric(18, 2))  # quantity in millions (M)
     
     # Lakebase Configuration
-    lakebase_cu = Column(Integer)
+    lakebase_cu = Column(Numeric(5, 1))
     lakebase_storage_gb = Column(Integer)
     lakebase_ha_nodes = Column(Integer)
     lakebase_backup_retention_days = Column(Integer)
